requirements.txt
groq==1.1.1
python-dotenv==1.2.2
Flask==3.1.3
requests==2.32.5